requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.0
torch>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
nest-asyncio>=1.6.0
sentence-transformers>=2.2.2
scikit-learn>=1.4.0
transformers>=4.38.0
faiss-cpu>=1.7.4
tqdm>=4.66.0
tenacity>=8.2.3
networkx>=3.1
jinja2>=3.1.2
aiohttp>=3.8.5
pytest>=7.4.0
pytest-asyncio>=0.21.1 